+++ CurtainCall/CurtainCall/Domain/UseCases/LoginUsecase.swift
@@ -10,7 +10,7 @@ import Combine
 
 protocol LoginUseCase {
     func loginWithApple(token: Data) -> AnyPublisher<String, Error>
-    func loginWithKakao() -> AnyPublisher<Void, Error>
+    func loginWithKakao(token: String) -> AnyPublisher<String, Error>
     func loginWithFacebook() -> AnyPublisher<Void, Error>
     func loginWithGoogle() -> AnyPublisher<Void, Error>
 }
@@ -32,10 +32,12 @@ final class LoginInteractor: LoginUseCase {
         
     }
     
-    func loginWithKakao() -> AnyPublisher<Void, Error> {
-        // TODO: 로그인 비즈니스 로직 수행 ex) API 호출
-        return Future<Void, Error> { promise in
-            promise(.success(print("카카오 로그인 성공했을 때, API 요청을 하거나 뭘 하자~")))
+    func loginWithKakao(token: String) -> AnyPublisher<String, Error> {
+        
+        // TODO: token값 서버로 전달
+        
+        return Future<String, Error> { promise in
+            promise(.success(token))
         }.eraseToAnyPublisher()
     }
     func loginWithFacebook() -> AnyPublisher<Void, Error> {
